DOKY-276 Migrate search functionality from Azure to Algolia (#205)
- Replaced Azure Search with Algolia for search and indexing services.
- Removed Azure-specific configurations, dependencies, and classes.
- Added support for Algolia-based indexing and search services.
- Updated `DocumentResultData` and `DocumentIndexData` to align with Algolia's schema.
- Adjusted application properties for Algolia configurations.
- Updated tests to reflect changes in indexing and search implementations.
- Enabled `java-test-fixtures` and added necessary dependencies to `build.gradle.kts` for test fixture support.

+++ src/main/kotlin/org/hkurh/doky/search/SearchResult.kt
@@ -27,5 +27,5 @@ package org.hkurh.doky.search
  */
 data class SearchResult(
     val documents: List<DocumentResultData>,
-    val totalCount: Long
+    val totalCount: Int
 )
